Database model define relation
