+++ api/infrastructure/database/database.go
@@ -5,7 +5,6 @@ import (
 	"os"
 	"strings"
 
-	"github.com/connthass/connthass/api/infrastructure/database/model"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
 )
@@ -23,7 +22,6 @@ func Init() {
 	if err != nil {
 		panic(err)
 	}
-
 	autoMigrate()
 }
 
@@ -48,47 +46,4 @@ func getConnectionString() string {
 // GetDB is called in models
 func GetDB() *gorm.DB {
 	return db
-}
-
-func autoMigrate() {
-	db.AutoMigrate(
-		model.User{},
-		model.Like{},
-		model.Category{},
-		model.Event{},
-		model.EventCategory{},
-		model.AppRole{},
-		model.SysRole{},
-		model.EntryEvent{},
-		model.Group{},
-		model.Member{},
-		model.CheckedEvent{},
-		model.Bookmark{},
-		model.Venue{},
-	)
-
-	db.Model(model.User{}).AddForeignKey("sys_role_id", "sys_roles(id)", "RESTRICT", "CASCADE")
-
-	db.Model(model.Bookmark{}).AddForeignKey("user_id", "users(id)", "CASCADE", "CASCADE")
-	db.Model(model.Bookmark{}).AddForeignKey("event_id", "events(id)", "CASCADE", "CASCADE")
-
-	db.Model(model.CheckedEvent{}).AddForeignKey("user_id", "users(id)", "CASCADE", "CASCADE")
-	db.Model(model.CheckedEvent{}).AddForeignKey("event_id", "events(id)", "CASCADE", "CASCADE")
-
-	db.Model(model.EntryEvent{}).AddForeignKey("user_id", "users(id)", "CASCADE", "CASCADE")
-	db.Model(model.EntryEvent{}).AddForeignKey("event_id", "events(id)", "CASCADE", "CASCADE")
-	db.Model(model.EntryEvent{}).AddForeignKey("app_role_id", "app_roles(id)", "RESTRICT", "CASCADE")
-
-	db.Model(model.Like{}).AddForeignKey("user_id", "users(id)", "CASCADE", "CASCADE")
-	db.Model(model.Like{}).AddForeignKey("category_id", "categories(id)", "CASCADE", "CASCADE")
-
-	db.Model(model.EventCategory{}).AddForeignKey("event_id", "events(id)", "CASCADE", "CASCADE")
-	db.Model(model.EventCategory{}).AddForeignKey("category_id", "categories(id)", "CASCADE", "CASCADE")
-
-	db.Model(model.Member{}).AddForeignKey("user_id", "users(id)", "CASCADE", "CASCADE")
-	db.Model(model.Member{}).AddForeignKey("app_role_id", "app_roles(id)", "RESTRICT", "CASCADE")
-	db.Model(model.Member{}).AddForeignKey("group_id", "`groups`(id)", "CASCADE", "CASCADE")
-
-	db.Model(model.Event{}).AddForeignKey("group_id", "`groups`(id)", "SET NULL", "CASCADE")
-	db.Model(model.Event{}).AddForeignKey("venue_id", "venues(id)", "SET NULL", "CASCADE")
 }
