+++ api/infrastructure/database/model/category.go
@@ -3,5 +3,7 @@ package model
 // Category 例）Go, React, もくもく
 type Category struct {
 	Base
-	Name string `gorm:"size:50;not null;unique_index"`
+	Name            string `gorm:"size:50;not null;unique_index"`
+	EventCategories []EventCategory
+	Likes           []Like
 }
