+++ api/infrastructure/database/model/event.go
@@ -6,7 +6,7 @@ import "time"
 type Event struct {
 	Base
 	BaseColor
-	title            string `gorm:"size:50;not null"`
+	Title            string `gorm:"size:50;not null"`
 	Description      string
 	Capacity         uint
 	ImageURL         string    `gorm:"size:255"`
@@ -16,5 +16,10 @@ type Event struct {
 	RecruitStartDate time.Time
 	RecruitEndDate   time.Time
 	GroupID          uint
-	VenueID          uint
+	VenueID          uint       `gorm:"not null"`
+	Group            Group      `gorm:"foreignkey:GroupID;association_foreignkey:Refer"`
+	Venue            Venue      `gorm:"foreignkey:VenueID;association_foreignkey:Refer"`
+	Bookmarks        []User     `gorm:"many2many:bookmarks`
+	Entries          []User     `gorm:"many2many:entry_events;"`
+	Categories       []Category `gorm:"many2many:event_categories;"`
 }
