+++ api/infrastructure/database/model/user.go
@@ -3,7 +3,11 @@ package model
 // User is connthass account
 type User struct {
 	Base
-	Name      string `gorm:"size:50;not null;unique_index"`
-	Nickname  string `gorm:"size:50;not null"`
-	SysRoleID uint
+	Name          string `gorm:"size:50;not null"`
+	UID           uint   `gorm:"not null;unique_index"`
+	SysRoleID     uint   `gorm:"not null"`
+	Bookmarks     []Bookmark
+	CheckedEvents []CheckedEvent
+	Likes         []Like
+	Members       []Member
 }
